refactor(Stadium): Field 클래스명 변경 -> Stadium
Stadium 디렉토리 하위에 Court 클래스 생성
