+++ src/main/java/com/example/tuktuk/domain/stadium/Court.java
@@ -1,0 +1,10 @@
+package com.example.tuktuk.domain.stadium;
+
+import java.util.List;
+
+public class Court {
+
+    CourtType courtType;
+    int hourlyRentFee;
+    List<byte[]> images;  // 이미지 데이터를 바이트 배열로 저장
+}
